pyproject.toml
[project]
name = "cyberpunk-looter"
version = "0.1.0"
description = "Discord bot with GraphRAG using Neo4j and memory system"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "discord.py>=2.4.0",
    "google-cloud-storage>=2.19.0",
    "langchain>=0.3.27",
    "neo4j-rust-ext>=5.28.2",
    "neo4j-graphrag[openai]>=1.10.0",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "anthropic>=0.71.0",
    "opentelemetry-distro>=0.59b0",
    "opentelemetry-exporter-otlp>=1.38.0",
    "pydantic>=2.12.2",
]

[dependency-groups]
dev = [
    "ruff>=0.14.1",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.1",
    "pytest-xdist>=3.3.1",
    "coverage>=7.3.0",
    "assertpy>=1.1",
    "pytest-env>=1.2.0",
    "pyright>=1.1.407",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--showlocals",
    "-ra",
    "--strict-markers",
    "--color=yes",
    "--durations=10",
]
markers = [
    "unit: Unit tests that don't require external dependencies",
    "integration: Integration tests that may require external services",
    "slow: Tests that take a long time to run",
    "asyncio: Tests that use async/await",
]
env = [
    "OPENAI_MODEL=fake",
    "OPENAI_BASE_URL=https://localhost:59998",
    "OPENAI_API_KEY=key3",
    "ANTHROPIC_API_KEY=key2",
    "OPENAI_EMBEDDINGS_KEY=key1",
    "OPENAI_EMBEDDINGS_BASE_URL=https://localhost:59999",
    "NEO4J_URI=bolt://localhost:7687",
    "NEO4J_USERNAME=neo4j",
    "NEO4J_PASSWORD=password",
    "GCS_BUCKET_NAME=bucket"
]

[tool.pyright]
include = ["src", "tests"]
exclude = ["**/__pycache__"]
defineConstant = { DEBUG = true }

reportMissingImports = "error"
reportMissingTypeStubs = false

[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true
